+++ src-tauri/src/frontend_types.rs
@@ -1,0 +1,9 @@
+#[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ChampionInfo {
+    pub url: String,
+    pub local_image: String,
+    pub win_rate: String,
+    pub pick_rate: String,
+    pub ban_rate: String,
+}
