+++ src-tauri/src/plugins/ugg/structs.rs
@@ -1,5 +1,3 @@
-use cached::proc_macro::cached;
-
 #[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Shards {
@@ -36,4 +34,45 @@ pub struct Data {
     pub role: String,
     pub rank: String,
     pub region: String,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct ItemsMap {
+    pub start: Vec<ItemValues>,
+    pub core: Vec<ItemValues>,
+    pub fourth: Vec<ItemValues>,
+    pub fifth: Vec<ItemValues>,
+    pub sixth: Vec<ItemValues>
+}
+
+#[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ItemValues {
+    pub name: String,
+    pub cost: String,
+    pub description: String,
+    pub local_image: String,
+    pub url: String,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct AbilitiesMap {
+    pub passive: Passive,
+    pub q: AbilitiesValue,
+    pub w: AbilitiesValue,
+    pub e: AbilitiesValue,
+    pub r: AbilitiesValue,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct AbilitiesValue {
+    pub image: String,
+    pub order: Vec<String>,
+    pub url: String,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct Passive {
+    pub image: String,
+    pub url: String,
 }
